add contrib to the package
