+++ elasticsearch/elasticsearch-5/src/main/java/com/hazelcast/jet/contrib/elasticsearch/ElasticsearchSinks.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.jet.elasticsearch;
+package com.hazelcast.jet.contrib.elasticsearch;
 
 import com.hazelcast.jet.function.ConsumerEx;
 import com.hazelcast.jet.function.FunctionEx;
@@ -35,7 +35,7 @@ import java.io.IOException;
 import static org.apache.http.auth.AuthScope.ANY;
 
 /**
- * Contains factory methods for Elasticsearch sinks.
+ * Contains factory methods for Elasticsearch sinks
  */
 public final class ElasticsearchSinks {
 
@@ -54,34 +54,35 @@ public final class ElasticsearchSinks {
      * @param destroyFn           called upon completion to release any resource
      */
     public static <T> Sink<T> elasticSearch(String name,
-                                            SupplierEx<RestHighLevelClient> clientSupplier,
+                                            SupplierEx<RestClient> clientSupplier,
                                             SupplierEx<BulkRequest> bulkRequestSupplier,
                                             FunctionEx<T, IndexRequest> indexFn,
-                                            ConsumerEx<RestHighLevelClient> destroyFn
+                                            ConsumerEx<RestClient> destroyFn
     ) {
         return SinkBuilder
                 .sinkBuilder("elasticSearch-" + name,
-                        ctx -> new BulkContext(clientSupplier.get(), bulkRequestSupplier, destroyFn))
+                        ctx -> new BulkContext(clientSupplier.get(), bulkRequestSupplier))
                 .<T>receiveFn((bulkContext, item) -> bulkContext.add(indexFn.apply(item)))
-                .flushFn(BulkContext::flush)
-                .destroyFn(BulkContext::close)
+                .flushFn(BulkContext::bulk)
+                .destroyFn(b -> destroyFn.accept(b.client))
                 .build();
     }
 
     /**
      * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx, FunctionEx, ConsumerEx)}
      * Creates a new {@link BulkRequest} with default options for each batch and
-     * closes the {@link RestHighLevelClient} upon completion.
+     * closes the {@link RestClient} upon completion.
      */
     public static <T> Sink<T> elasticSearch(String name,
-                                            SupplierEx<RestHighLevelClient> clientSupplier,
+                                            SupplierEx<RestClient> clientSupplier,
                                             FunctionEx<T, IndexRequest> indexFn
     ) {
-        return elasticSearch(name, clientSupplier, BulkRequest::new, indexFn, RestHighLevelClient::close);
+        return elasticSearch(name, clientSupplier, BulkRequest::new, indexFn, RestClient::close);
     }
 
+
     /**
-     * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx, FunctionEx, ConsumerEx)}
+     * Convenience for {@link #elasticSearch(String, SupplierEx, FunctionEx)}
      * Rest client is configured with basic authentication.
      */
     public static <T> Sink<T> elasticSearch(String name,
@@ -92,30 +93,27 @@ public final class ElasticsearchSinks {
         return elasticSearch(name, () -> buildClient(username, password, hostname, port), indexFn);
     }
 
-    static RestHighLevelClient buildClient(String username, String password, String hostname, int port) {
+    static RestClient buildClient(String username, String password, String hostname, int port) {
         CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
         credentialsProvider.setCredentials(ANY, new UsernamePasswordCredentials(username, password));
-        return new RestHighLevelClient(
-                RestClient.builder(new HttpHost(hostname, port))
-                          .setHttpClientConfigCallback(httpClientBuilder ->
-                                  httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider))
-        );
+
+        return RestClient.builder(new HttpHost(hostname, port))
+                         .setHttpClientConfigCallback(httpClientBuilder ->
+                                 httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider)).build();
     }
 
     private static final class BulkContext {
 
-        private final RestHighLevelClient client;
+        private final RestClient client;
+        private final RestHighLevelClient highLevelClient;
         private final SupplierEx<BulkRequest> bulkRequestSupplier;
-        private final ConsumerEx<RestHighLevelClient> destroyFn;
 
         private BulkRequest bulkRequest;
 
-        private BulkContext(RestHighLevelClient client, SupplierEx<BulkRequest> bulkRequestSupplier,
-                            ConsumerEx<RestHighLevelClient> destroyFn) {
+        private BulkContext(RestClient client, SupplierEx<BulkRequest> bulkRequestSupplier) {
             this.client = client;
+            this.highLevelClient = new RestHighLevelClient(client);
             this.bulkRequestSupplier = bulkRequestSupplier;
-            this.destroyFn = destroyFn;
-
             this.bulkRequest = bulkRequestSupplier.get();
         }
 
@@ -123,13 +121,9 @@ public final class ElasticsearchSinks {
             bulkRequest.add(request);
         }
 
-        private void flush() throws IOException {
-            client.bulk(bulkRequest);
+        private void bulk() throws IOException {
+            highLevelClient.bulk(bulkRequest);
             bulkRequest = bulkRequestSupplier.get();
-        }
-
-        private void close() {
-            destroyFn.accept(client);
         }
     }
