+++ elasticsearch/elasticsearch-6/src/main/java/com/hazelcast/jet/contrib/elasticsearch/ElasticsearchSinks.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.jet.elasticsearch;
+package com.hazelcast.jet.contrib.elasticsearch;
 
 import com.hazelcast.jet.function.ConsumerEx;
 import com.hazelcast.jet.function.FunctionEx;
@@ -25,10 +25,8 @@ import org.apache.http.HttpHost;
 import org.apache.http.auth.UsernamePasswordCredentials;
 import org.apache.http.client.CredentialsProvider;
 import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.elasticsearch.action.ActionRequest;
 import org.elasticsearch.action.bulk.BulkRequest;
 import org.elasticsearch.action.index.IndexRequest;
-import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestClient;
 import org.elasticsearch.client.RestHighLevelClient;
 
@@ -37,7 +35,7 @@ import java.io.IOException;
 import static org.apache.http.auth.AuthScope.ANY;
 
 /**
- * Contains factory methods for Elasticsearch sinks
+ * Contains factory methods for Elasticsearch sinks.
  */
 public final class ElasticsearchSinks {
 
@@ -53,41 +51,37 @@ public final class ElasticsearchSinks {
      * @param bulkRequestSupplier Bulk request supplier, will be called to obtain a
      *                            new {@link BulkRequest} instance after each call.
      * @param indexFn             Creates an {@link IndexRequest} for each object
-     * @param optionsFn           obtains a {@link RequestOptions} for each request
      * @param destroyFn           called upon completion to release any resource
      */
     public static <T> Sink<T> elasticSearch(String name,
                                             SupplierEx<RestHighLevelClient> clientSupplier,
                                             SupplierEx<BulkRequest> bulkRequestSupplier,
                                             FunctionEx<T, IndexRequest> indexFn,
-                                            FunctionEx<? super ActionRequest, RequestOptions> optionsFn,
                                             ConsumerEx<RestHighLevelClient> destroyFn
     ) {
         return SinkBuilder
                 .sinkBuilder("elasticSearch-" + name,
-                        ctx -> new BulkContext(clientSupplier.get(), bulkRequestSupplier, optionsFn, destroyFn))
+                        ctx -> new BulkContext(clientSupplier.get(), bulkRequestSupplier, destroyFn))
                 .<T>receiveFn((bulkContext, item) -> bulkContext.add(indexFn.apply(item)))
-                .flushFn(BulkContext::bulk)
+                .flushFn(BulkContext::flush)
                 .destroyFn(BulkContext::close)
                 .build();
     }
 
     /**
-     * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx, FunctionEx, FunctionEx, ConsumerEx)}
-     * Uses {@link RequestOptions#DEFAULT}, creates a new {@link BulkRequest}
-     * with default options for each batch and closes the {@link
-     * RestHighLevelClient} upon completion.
+     * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx, FunctionEx, ConsumerEx)}
+     * Creates a new {@link BulkRequest} with default options for each batch and
+     * closes the {@link RestHighLevelClient} upon completion.
      */
     public static <T> Sink<T> elasticSearch(String name,
                                             SupplierEx<RestHighLevelClient> clientSupplier,
                                             FunctionEx<T, IndexRequest> indexFn
     ) {
-        return elasticSearch(name, clientSupplier, BulkRequest::new, indexFn,
-                request -> RequestOptions.DEFAULT, RestHighLevelClient::close);
+        return elasticSearch(name, clientSupplier, BulkRequest::new, indexFn, RestHighLevelClient::close);
     }
 
     /**
-     * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx, FunctionEx, FunctionEx, ConsumerEx)}
+     * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx, FunctionEx, ConsumerEx)}
      * Rest client is configured with basic authentication.
      */
     public static <T> Sink<T> elasticSearch(String name,
@@ -112,17 +106,14 @@ public final class ElasticsearchSinks {
 
         private final RestHighLevelClient client;
         private final SupplierEx<BulkRequest> bulkRequestSupplier;
-        private final FunctionEx<? super ActionRequest, RequestOptions> optionsFn;
         private final ConsumerEx<RestHighLevelClient> destroyFn;
 
         private BulkRequest bulkRequest;
 
         private BulkContext(RestHighLevelClient client, SupplierEx<BulkRequest> bulkRequestSupplier,
-                            FunctionEx<? super ActionRequest, RequestOptions> optionsFn,
                             ConsumerEx<RestHighLevelClient> destroyFn) {
             this.client = client;
             this.bulkRequestSupplier = bulkRequestSupplier;
-            this.optionsFn = optionsFn;
             this.destroyFn = destroyFn;
 
             this.bulkRequest = bulkRequestSupplier.get();
@@ -132,8 +123,8 @@ public final class ElasticsearchSinks {
             bulkRequest.add(request);
         }
 
-        private void bulk() throws IOException {
-            client.bulk(bulkRequest, optionsFn.apply(bulkRequest));
+        private void flush() throws IOException {
+            client.bulk(bulkRequest);
             bulkRequest = bulkRequestSupplier.get();
         }
