+++ elasticsearch/elasticsearch-7/src/main/java/com/hazelcast/jet/contrib/elasticsearch/ElasticsearchSources.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.jet.elasticsearch;
+package com.hazelcast.jet.contrib.elasticsearch;
 
 import com.hazelcast.jet.function.ConsumerEx;
 import com.hazelcast.jet.function.FunctionEx;
@@ -22,17 +22,16 @@ import com.hazelcast.jet.function.SupplierEx;
 import com.hazelcast.jet.pipeline.BatchSource;
 import com.hazelcast.jet.pipeline.SourceBuilder;
 import com.hazelcast.jet.pipeline.SourceBuilder.SourceBuffer;
+import org.elasticsearch.action.ActionRequest;
 import org.elasticsearch.action.search.ClearScrollRequest;
 import org.elasticsearch.action.search.SearchRequest;
 import org.elasticsearch.action.search.SearchResponse;
 import org.elasticsearch.action.search.SearchScrollRequest;
-import org.elasticsearch.client.RestClient;
+import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.elasticsearch.search.SearchHit;
 
 import java.io.IOException;
-
-import static com.hazelcast.jet.elasticsearch.ElasticsearchSinks.buildClient;
 
 /**
  * Contains factory methods for Elasticsearch sources
@@ -53,40 +52,42 @@ public final class ElasticsearchSources {
      * @param searchRequestSupplier Search request supplier
      * @param scrollTimeout         scroll keep alive time
      * @param hitMapperFn           maps search hits to output items
+     * @param optionsFn             obtains a {@link RequestOptions} for each request
      * @param destroyFn             called upon completion to release any resource
      * @param <T>                   type of items emitted downstream
      */
     public static <T> BatchSource<T> elasticSearch(String name,
-                                                   SupplierEx<RestClient> clientSupplier,
+                                                   SupplierEx<RestHighLevelClient> clientSupplier,
                                                    SupplierEx<SearchRequest> searchRequestSupplier,
                                                    String scrollTimeout,
                                                    FunctionEx<SearchHit, T> hitMapperFn,
-                                                   ConsumerEx<RestClient> destroyFn
+                                                   FunctionEx<? super ActionRequest, RequestOptions> optionsFn,
+                                                   ConsumerEx<RestHighLevelClient> destroyFn
+
     ) {
         return SourceBuilder
-                .batch(name, ctx -> {
-                    RestClient client = clientSupplier.get();
-                    SearchRequest searchRequest = searchRequestSupplier.get();
-                    return new SearchContext<>(client, scrollTimeout, hitMapperFn, searchRequest, destroyFn);
-                })
+                .batch(name, ctx -> new SearchContext<>(clientSupplier.get(), scrollTimeout,
+                        hitMapperFn, searchRequestSupplier.get(), optionsFn, destroyFn))
                 .<T>fillBufferFn(SearchContext::fillBuffer)
                 .destroyFn(SearchContext::close)
                 .build();
     }
 
     /**
-     * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx, String, FunctionEx, ConsumerEx)}.
-     * Uses {@link #DEFAULT_SCROLL_TIMEOUT} for scroll timeout, emits string
-     * representation of items using {@link SearchHit#getSourceAsString()} and
-     * closes the {@link RestClient} upon completion.
+     * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx, String, FunctionEx, FunctionEx, ConsumerEx)}.
+     * Uses {@link #DEFAULT_SCROLL_TIMEOUT} for scroll timeout and {@link
+     * RequestOptions#DEFAULT}, emits string representation of items using
+     * {@link SearchHit#getSourceAsString()} and closes the {@link
+     * RestHighLevelClient} upon completion.
      */
     public static BatchSource<String> elasticSearch(String name,
-                                                    SupplierEx<RestClient> clientSupplier,
+                                                    SupplierEx<RestHighLevelClient> clientSupplier,
                                                     SupplierEx<SearchRequest> searchRequestSupplier
     ) {
-        return elasticSearch(name, clientSupplier, searchRequestSupplier,
-                DEFAULT_SCROLL_TIMEOUT, SearchHit::getSourceAsString, RestClient::close);
+        return elasticSearch(name, clientSupplier, searchRequestSupplier, DEFAULT_SCROLL_TIMEOUT,
+                SearchHit::getSourceAsString, request -> RequestOptions.DEFAULT, RestHighLevelClient::close);
     }
+
 
     /**
      * Convenience for {@link #elasticSearch(String, SupplierEx, SupplierEx)}.
@@ -97,29 +98,33 @@ public final class ElasticsearchSources {
                                                     String hostname, int port,
                                                     SupplierEx<SearchRequest> searchRequestSupplier
     ) {
-        return elasticSearch(name, () -> buildClient(username, password, hostname, port), searchRequestSupplier);
+        return elasticSearch(name, () -> ElasticsearchSinks.buildClient(username, password, hostname, port),
+                searchRequestSupplier);
     }
 
     private static final class SearchContext<T> {
 
-        private final RestClient client;
-        private final RestHighLevelClient highLevelClient;
+        private final RestHighLevelClient client;
         private final String scrollInterval;
         private final FunctionEx<SearchHit, T> hitMapperFn;
-        private final ConsumerEx<RestClient> destroyFn;
+        private final FunctionEx<? super ActionRequest, RequestOptions> optionsFn;
+        private final ConsumerEx<RestHighLevelClient> destroyFn;
 
         private SearchResponse searchResponse;
 
-        private SearchContext(RestClient client, String scrollInterval, FunctionEx<SearchHit, T> hitMapperFn,
-                              SearchRequest searchRequest, ConsumerEx<RestClient> destroyFn) throws IOException {
+        private SearchContext(RestHighLevelClient client, String scrollInterval,
+                              FunctionEx<SearchHit, T> hitMapperFn, SearchRequest searchRequest,
+                              FunctionEx<? super ActionRequest, RequestOptions> optionsFn,
+                              ConsumerEx<RestHighLevelClient> destroyFn
+        ) throws IOException {
             this.client = client;
-            this.highLevelClient = new RestHighLevelClient(client);
             this.scrollInterval = scrollInterval;
             this.hitMapperFn = hitMapperFn;
+            this.optionsFn = optionsFn;
             this.destroyFn = destroyFn;
 
             searchRequest.scroll(scrollInterval);
-            searchResponse = highLevelClient.search(searchRequest);
+            searchResponse = client.search(searchRequest, optionsFn.apply(searchRequest));
         }
 
         private void fillBuffer(SourceBuffer<T> buffer) throws IOException {
@@ -137,13 +142,13 @@ public final class ElasticsearchSources {
 
             SearchScrollRequest scrollRequest = new SearchScrollRequest(searchResponse.getScrollId());
             scrollRequest.scroll(scrollInterval);
-            searchResponse = highLevelClient.searchScroll(scrollRequest);
+            searchResponse = client.scroll(scrollRequest, optionsFn.apply(scrollRequest));
         }
 
         private void clearScroll() throws IOException {
             ClearScrollRequest clearScrollRequest = new ClearScrollRequest();
             clearScrollRequest.addScrollId(searchResponse.getScrollId());
-            highLevelClient.clearScroll(clearScrollRequest);
+            client.clearScroll(clearScrollRequest, RequestOptions.DEFAULT);
         }
 
         private void close() throws IOException {
