+++ elasticsearch/elasticsearch-7/src/test/java/com/hazelcast/jet/contrib/elasticsearch/ElasticsearchSinkTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.jet.elasticsearch;
+package com.hazelcast.jet.contrib.elasticsearch;
 
 import com.hazelcast.jet.pipeline.Pipeline;
 import com.hazelcast.jet.pipeline.Sources;
@@ -22,18 +22,22 @@ import org.junit.Test;
 
 import java.io.IOException;
 
+import static com.hazelcast.jet.contrib.elasticsearch.ElasticsearchSinks.elasticSearch;
+
 public class ElasticsearchSinkTest extends ElasticsearchBaseTest {
 
     @Test
     public void test_elasticSearchSink() throws IOException {
+
         String containerAddress = container.getHttpHostAddress();
 
         Pipeline p = Pipeline.create();
         p.drawFrom(Sources.list(userList))
-         .drainTo(ElasticsearchSinks.elasticSearch(indexName, () -> createClient(containerAddress), indexFn(indexName)));
+         .drainTo(elasticSearch(indexName, () -> createClient(containerAddress), indexFn(indexName)));
 
         jet.newJob(p).join();
 
         assertIndexes();
     }
+
 }
